Inputs + Common: implements input handling for wheel and panel from old bottom shell
This commit implements many of the inputs that were previously handled in the old bottom shell repository. It implements the brake switch, forward/reverse, and whether the message came from the wheel or the switch panel. It also stores information about the cruise regen.

This also implements a templated utility function that can parse a CAN message buffer into a compile-time specified data type at a compile-time specified offset. This removes the need for modifying the FlexCAN_T4 library while providing us similar levels of functionality.